refactor(engine): split interactive, replicas, and project_label out of mod.rs
